modules/platforms/ignite-python-thin-client/requirements/install.txt
# these pip packages are necessary for the pyignite to run

attrs>=20.3.0
